+++ web/src/components/select_course/tea_sel_course.vue
@@ -1,0 +1,13 @@
+<template>
+    
+</template>
+
+<script>
+    export default {
+        name: "tea_sel_course"
+    }
+</script>
+
+<style scoped>
+
+</style>
